+++ backend/application-modules/api-server/src/main/java/com/softeer/config/WebConfiguration.java
@@ -1,11 +1,23 @@
-package com.softeer.security;
+package com.softeer.config;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+import java.util.List;
+
 @Configuration
-public class SecurityConfiguration implements WebMvcConfigurer {
+@RequiredArgsConstructor
+public class WebConfiguration implements WebMvcConfigurer {
+
+    private final LoginUserIdArgumentResolver loginUserIdArgumentResolver;
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(loginUserIdArgumentResolver);
+    }
 
     @Override
     public void addCorsMappings(CorsRegistry registry) {
