+++ backend/application-modules/api-server/src/main/java/com/softeer/config/auth/JwtGenerator.java
@@ -1,4 +1,4 @@
-package com.softeer.security.auth;
+package com.softeer.config.auth;
 
 import com.softeer.entity.Role;
 import io.jsonwebtoken.Jwts;
