+++ backend/application-modules/api-server/src/main/java/com/softeer/config/auth/Token.java
@@ -1,3 +1,3 @@
-package com.softeer.security.auth;
+package com.softeer.config.auth;
 
 public record Token(String value) { }
